requirements.txt
black==23.3.0
coverage==7.2.3
Django==4.2
mock==5.0.1
Pillow==9.5.0
pyserial==3.5
